requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2
reportlab==4.2.5
weasyprint==63.1
razorpay==1.4.2
sentry-sdk==2.16.0
requests==2.32.3
